@@ -5,7 +5,7 @@
     <parent>
         <artifactId>s3-server</artifactId>
         <groupId>de.mindconsulting</groupId>
-        <version>2.3.2</version>
+        <version>2.5.0</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
     <artifactId>ladon-s3-server-jetty</artifactId>
@@ -13,9 +13,9 @@
     <description>Ladon S3 Jetty Example</description>
 
     <properties>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <jetty.version>9.4.51.v20230217</jetty.version>
+        <maven.compiler.source>21</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
+        <jetty.version>12.1.0</jetty.version>
     </properties>
     <dependencies>
         <dependency>
@@ -26,13 +26,8 @@
         <dependency>
             <groupId>de.mindconsulting</groupId>
             <artifactId>ladon-s3-server-fs</artifactId>
-            <version>2.3.2</version>
+            <version>2.4.1</version>
             <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>javax.activation</groupId>
-            <artifactId>activation</artifactId>
-            <version>1.1.1</version>
         </dependency>
         <dependency>
             <groupId>org.eclipse.jetty</groupId>
@@ -40,14 +35,14 @@
             <version>${jetty.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.eclipse.jetty</groupId>
-            <artifactId>jetty-servlet</artifactId>
+            <groupId>org.eclipse.jetty.ee11</groupId>
+            <artifactId>jetty-ee11-servlet</artifactId>
             <version>${jetty.version}</version>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-simple</artifactId>
-            <version>1.7.32</version>
+            <version>2.0.16</version>
         </dependency>
     </dependencies>
     <developers>
